requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cpf==0.1.0
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2==2.9.4
pytz==2022.5
sqlparse==0.4.3
django-rest-swagger==2.2.0
swagger-spec-validator==3.0.2
drf-yasg==1.21.4
django-rest-auth
djangorestframework-simplejwt